[CBRD-25216] Check if there is enough buffer to store the user-specified name (#4946)
http://jira.cubrid.org/browse/CBRD-25216

We found that the buffer of the pt_get_name_without_current_user_name function may be insufficient to copy the user-specified name. Also, we found that if the name of the auto-increment serial exceeds 254 bytes, it is truncated to 254 bytes when the owner of the table changes. In debug mode, if the table name exceeds 222 bytes, an assertion occurs in sm_user_specified_name. To prevent this, there are the following changes:

1. Copy only the qualifier name from the user-specified name in the pt_get_name_without_current_user_name function.
2. Add sm_user_specified_name_for_serial function.
3. Check if there is enough buffer in sm_user_specified_name and sm_qualifier_name functions.
4. Do not truncate the length of the new serial name to 254 bytes in the au_change_serial_owner function.
5. Add a safe guard to sm_user_specified_name and sm_user_specified_name_for_serial.