[CUBRIDQA-1211] Modify Jenkinsfile for manual test automation (#4909)
http://jira.cubrid.org/browse/CUBRIDQA-1211

In the 'Linux Release' and 'Linux Debug' stages, I added a script block to conditionally skip testing for feature branches. This logic checks if the branch name matches the pattern ^feature/.* before executing the testing steps.
In the 'Windows Release' stage, I added a when block to conditionally skip the entire stage for feature branches. This ensures that the 'Windows Release' stage is skipped for branches matching the specified pattern.
In the final post block, I used an if-else condition within the script block to trigger different deployment jobs based on the branch name. If the branch is not a feature branch, it triggers ${DEPLOY_JOB}, otherwise ${DEPLOY_JOB_FOR_MANUAL} is triggered.