[CBRD-25285] Add a new ClassLoader to avoid reloading classes that load JNI in a Java SP Server (#5076) (#5090)
http://jira.cubrid.org/browse/CBRD-25285

A new classloader (StoredProcedureStaticClassLoader) will be added to avoid reloading classes that load JNI. By loadjava's -j option, classes are copied to /java_static directory and the classloader will try to find classes in the path.
The classloader does not check whether any file is modified in the path, Also is not be re-initialized like StoredProcedureClassLoader. To update overwritten classes in /java_static, restarting Java SP server(cub_javasp) is required.