[CBRD-25296] Assertion fail on a DATETIMELTZ value to a SP's parameter of DATETIME type (#5099)
http://jira.cubrid.org/browse/CBRD-25296

When calling a function or procedure, it's necessary to verify that the types of the provided arguments are supported.
These arguments are passed as an array of DB_VALUE, and a newly introduced function named is_supported_dbtype() is responsible for checking their DB_TYPE.
In situations where the argument type is not supported, an error message should be generated in the following format: 'unsupported argument type - <DB_TYPE>', where the <DB_TYPE> placeholder is replaced with the actual string representation of the DB_TYPE.

To print DB_TYPE to a string representation, a new function called describe_type() has been added to the db_value_printer module. This addition was necessary because the existing compat layer does not offer such functionality, although a similar function, qdump_data_type_string(), exists within query_dump.c.