[LETS-166] refactor redo_job_impl as a non-template class (#2929)
http://jira.cubrid.org/browse/LETS-166

Profiling showed that the overhead incurred by the allocator constantly constructing (and, probably, destructing) a large amount of objects (in particular `redo_job_impl` instances) is considerable.

Reimplement the `redo_job_impl` as a non-template class. This is needed for further refactoring when `redo_job_impl` will be used in a reusable manner (sic!) in order to minimize the overhead incurred by the memory allocator constantly constructing/destructing a large number of such objects.