[CBRD-25376] revise error message that may occur when loading ha_node_list and ha_replica_list
http://jira.cubrid.org/browse/CBRD-25376

**Description**

If incorrect information is used in ha_node_list or ha_replica_list, the following error message is displayed in master.err, and it is hard to find the cause of the error based on this message and fix this error.
So, it need to be modified so that the error message is output differently depending on the error type.
Please refer to the issue for the changed error message.