[CBRD-25316] For UPDATE JOIN statements, aggregate functions or analytic functions cannot be used in the SET and WHERE clauses (#5143)
http://jira.cubrid.org/browse/CBRD-25316

The execution result of the UPDATE JOIN statement depends on whether the USE_MERGE hint is used. To prevent this issue, do not allow the use of analytic functions in the SET and WHERE clauses of UPDATE JOIN statements. Aggregate functions could not be used in UPDATE statements, even if they were not UPDATE JOIN statements.