[CBRD-25396] When using GROUP BY and ORDER BY simultaneously within a subquery, the NO_MERGE hint does not function correctly. (#5244)
http://jira.cubrid.org/browse/CBRD-25396

If the NO_MERGE hint moves to the derived subquery, it will not affect the agg_sel subquery, so the NO_MERGE hint is not moved to the derived subquery.
Additionally, if the subquery has the QUERY_CACHE hint, it should not be merged, so it is treated together with the NO_MERGE hint. 
All hints except for NO_MERGE and QUERY_CACHE are moved to the derived subquery.