[CBRD-25323] Remove unnecessary code (#5404)
http://jira.cubrid.org/browse/CBRD-25323

This `assert(false);` was added to later verify the behavior difference between non-ANSI OUTER JOIN and ANSI OUTER JOIN syntax. It should have been removed before the commit, but this step was missed. Since we've decided to accept the differences between the two join syntaxes, we are now removing this code.