[CBRD-25532] Remove the unnecessary archiving log volume when backupdb (#5432)
http://jira.cubrid.org/browse/CBRD-25532

Purpose

To consider active transactions present at the time of backup as restore targets, the logs in memory are flushed during the backup, and the active log volume is archived. However, since active transactions are subject to UNDO at the time of restore, creating unnecessary archive log volumes is not required. Therefore, this issue involves removing the creation of unnecessary archive log volumes.