[CBRD-25540] If there is a stored procedure in query_spec, [user_schema.] is not deleted but unloaded. (#5442)
http://jira.cubrid.org/browse/CBRD-25540

This PR can only be tested when PL/CSQL phase-1 is merged. (Related to adding user_schema to SP)

When removing [user_schema.] from the query_spec of a view, check and remove the flag (PT_PRINT_NO_CURRENT_USER_NAME) in the rewrite statement.
However, pt_print_function() doesn't have logic to check for flags, so we added logic to check for flags.