[CBRD-25050] Add check for log record header corruption (#5469)
http://jira.cubrid.org/browse/CBRD-25050

Purpose

If the system terminates abnormally during the flushing of a log record, it cannot be guaranteed that the log record on disk is intact. Currently, during the process of specifying the recovery range, the system sequentially scans the log records to detect any corruption. However, it does not check whether the forw_lsa stored in the log record contains invalid values. As a result, issues may arise when accessing a log page or reading an incorrect log record due to an invalid forw_lsa value.  Therefore, we will resolve this issue by adding a check for the validity of the forw_lsa value.