[CBRD-25599] Modify the assert condition to exclude system transactions from the check in recovery redo phase (#5518)
http://jira.cubrid.org/browse/CBRD-25599

Purpose

In the analysis phase, completed transactions are removed from the transaction table.
During the redo phase, when encountering a LOG_COMMIT or LOG_ABORT log, the system checks whether the corresponding transaction information exists in the transaction table.
If the transaction is found in the table, it indicates an issue that occurred during the analysis phase. 
However, the current assertion does not take system transactions into account. Since system transactions are not cleaned up during the analysis phase, the assertion needs to be modified to exclude checks for system transactions.