[CBRD-25507] Slip: Handling a cursor reaches to the end of result set (#5526)
http://jira.cubrid.org/browse/CBRD-25507

This change adds handling for cases in the server-side JDBC fetch() protocol where the server-side cursor reaches the end (S_END) and no tuples are returned.

In this case, the number of tuples returned in fetchInfo (numFetched) is set to 0, and tuples is set to null since no records were returned. The regression mentioned in the issue's comment was caused by accessing the tuples in this case, leading to a NullPointerException.