[CBRD-25480] Backport from develop to 11.2 - When an error occurs in loaddb, it does not stop and continues execution. (#5974)
http://jira.cubrid.org/browse/CBRD-25480

If an error occurs while running loaddb, loaddb should be stopped immediately and an error message should be output. However, in certain cases, even if an error occurs, no error message is output and loaddb terminates as if it was processed normally.
And the part where the line number was missing or displayed incorrectly was corrected.