[CUBVEC-96] Fix COSINE metric issues with all-zero vectors (#6291)
http://jira.cubrid.org/browse/CUBVEC-96

Eliminate issues where undefined behavior occurs when an all-zero vector is used in index operations or distance computations with the cosine metric. 
1. prevent insertion of all-zero vectors into the index, and ensure they do not produce results when used in search conditions. 
2. return NULL for cosine distance calculations and distance functions when an all-zero vector is involved.

In hnsw_add_element and hnsw_search_element, immediately return when the metric is cosine and the vector is all-zero.
In vector_cosine_distance, check for all-zero vectors in the arguments and return NULL if detected.