[CBRD-26390] Avoid error and treat as NO-OP when a non-owner grantor attempts to GRANT an owner’s class while running loaddb with --no-user-specified-name (#6654)
http://jira.cubrid.org/browse/CBRD-26390

When loading a schema unloaded from version 11.0v or earlier, 
GRANT statements that re-grant an owner’s class to the owner are now treated as a NO-OP when using the --no-user-specified-name option.

However, when executed through CSQL or by users directly, 
the behavior remains unchanged and continues to raise the error "Cannot issue GRANT/REVOKE to owner of a class.