[CBRD-26390] (backport) Avoid error and treat as NO-OP when a non-owner grantor attempts to GRANT an owner’s class while running loaddb with --no-user-specified-name (#6657)
http://jira.cubrid.org/browse/CBRD-26390

When loading a schema unloaded from version 11.0v or earlier, 
GRANT statements that re-grant an owner’s class to the owner are now treated as a NO-OP when using the --no-user-specified-name option.

However, when executed through CSQL or by users directly, 
the behavior remains unchanged and continues to raise the error "Cannot issue GRANT/REVOKE to owner of a class.

In addition, propagate_revoke() is now skipped by returning NO_ERROR immediately for the following reasons.
1. In versions 11.3 and earlier, propagate_revoke (cascade revoke) does not operate correctly.
This issue was fixed starting from version 11.4.

2. In debug mode, when the grantor is different from the owner and more than one grantor exists,
executing a REVOKE triggers propagate_revoke, which leads to an assertion failure inside
au_propagate_del_new_auth(), resulting in a core dump.